requirements.txt
Django>=4.2
djangorestframework
djangorestframework-simplejwt
psycopg2-binary
celery
redis
django-cors-headers
geopy
sentence-transformers
numpy
scikit-learn